Banyak yang diubah. Kyk, beberapa method Human gw pindahin ke private, constructor Human gw agak ubah", beberapa variabel gw ganti jadi float supaya bisa "dynamic", exp dan gold gw tambahin juga, ada expRequirement di class Human, item sekarang ada harga!, dan class Monster gw tambahin.

+++ Diablo1/Diablo1/Base.h
@@ -6,24 +6,25 @@
 class Base
 {
 protected:
-	int damage;
-	int chanceToHit;
-	int evade;
-	int speed;
-	int maxHealth;
-	int maxStamina;
+	// semua yang secondary stats, kecuali armor, adalah floating point
+	float damage;
+	float chanceToHit;
+	float evade;
+	float speed;
+	float maxHealth;
+	float maxStamina;
 	int armor;
 	int level;
 	char name[50];
 
 public:
 	//getter
-	int getDamage() const { return damage; }
-	int getChanceToHit() const { return chanceToHit; }
-	int getEvade() const { return evade; }
-	int getSpeed() const { return speed; }
-	int getMaxHealth() const { return maxHealth; }
-	int getMaxStamina() const { return maxStamina; }
+	float getDamage() const { return damage; }
+	float getChanceToHit() const { return chanceToHit; }
+	float getEvade() const { return evade; }
+	float getSpeed() const { return speed; }
+	float getMaxHealth() const { return maxHealth; }
+	float getMaxStamina() const { return maxStamina; }
 	int getArmor() const { return armor; }
 	int getLevel() const { return level; }
 	const char* getName() const { return name; }
